These are type cleanups and a genuine bug fix. Signed-off-by: Ronald G. Minnich <rminnich@gmail.com> Acked-by: Stefan Reinauer <stepan@coresystems.de>
git-svn-id: svn://coreboot.org/repository/LinuxBIOSv3@494 f3766cd6-281f-0410-b1cd-43a5c92072e9

+++ device/pci_device.c
@@ -661,7 +661,7 @@ void pci_dev_init(struct device *dev)
 	ram = pci_rom_load(dev, rom);
 	if (ram == NULL)
 		return;
-	run_bios(dev, ram);
+	run_bios(dev, (unsigned long)ram);
 #endif
 }
